requirements.txt
kaldi-io
kaldi
numpy==1.16.2
pytorch==1.1.0
Python==3.6.8
